Rename: snake_case -> camelCase
I've been debating this for years: "Which is better for variables and
properties in AHK? snake_case or camelCase?" Ultimately, one thing
decided it for me: Built-in properties don't have an underscore between
words. So, to stay consistent with AHK's syntax, I've made the switch
from snake_case to camelCase.

This is what I was hinting at in
4698dfcbecf9ee06c024a8dc1ba1738c9b7de139.