enh: create Custom Layer
This is what I was working on on and off
for the 7 months I wasn't committing.
I wasn't committing for the longest time
because this Custom Layer would get its sent modifiers stuck,
and for 7 months, I couldn't figure out why.

From my perspective, it was just down to bad timing.
When I worked on the custom layer in August 2023,
there was a bug in AHK that I was unaware of
that would only be fixed in September.

My comments in default-settings and launcher explain it:
"Before v2.0.7, hook hotkeys were not recognizing modifiers
which were pressed down by SendInput.
Custom Layer would not work in v2.0.6 and below."

So in August 2023, I used SendEvent in place of SendInput,
but I've now concluded that SendEvent
is too unreliable to be worth using at any capacity.
It also get the modifiers stuck.

In January 2024, I finally discovered the bug.
I was annoyed that it was a bug with AHK itself all along,
but I was satisified that it was finally over.

scope:
    default-settings.ahk
    launcher.ahk
    keyboard.ahk
    # Keys
    ## Backspace
    ## Modifiers
    ### Custom Layer
    # Functions