enh: remap most mouse shortcuts
Yes, this is quite the commit.
Surprisingly, this was not 7 months in the making but 3 weeks.

The reason I've made such a huge overhaul?
I have a new philosophy of not interfering as much as possible
with native shortcuts and functions.

My old mouse.ahk did nothing when RButton and MButton were down.
But the native down events do have a purpose!
For example, in VS Code, holding down MButton will make a box selection,
but I didn't know that existed
because I was always blocking it.

The rationale behind this philosophy
is that I don't want to lose useful functionality,
and as much as possible, I want to keep native functions as they are
so that when I use a computer that's not mine,
it won't be as jarring.

I have achieved this by making each shortcut
either based on XButton1 or XButton2.
Then I realized I could theme the shortcuts.
So XButton1 is for window and typing shortcuts
while XButton2 is for tab shortcuts.

I've also added many new shortcuts
while making all the shortcuts as a whole easier to memorize
(because they are themed and related shortcuts are similar).

The newest functionality
is sending keys such as Enter and shortcuts like Ctrl+Z.
It's crazy how many functions can be crammed onto the mouse.

scope:
    mouse.ahk
    classes.ahk
    functions.ahk
    # Window
    keyboard.ahk
    # Specific App