enh: remove C_Timer
I'm not sure since I haven't used it in literally years,
but I think it was supposed to allow me to
check if a timer was running.

Not sure about this again,
but I think I never used it lol
and just made it because I made C_KeyWait
and had the idea that I could do the same for SetTimer.

I don't think it's actually useful,
so I'm removing it.